+++ lib/at_ex.ex
@@ -23,8 +23,7 @@ defmodule AtEx do
   """
   alias AtEx.Gateway.{
     Airtime,
-    Application,
-    Sms
+    Application
   }
 
   @doc """
@@ -69,55 +68,4 @@ defmodule AtEx do
       {:ok, %{"UserData" => %{"balance" => "ZAR -1.4572"}}}
   """
   defdelegate get_application_data(), to: Application, as: :get_data
-
-  @doc """
-  This function builds and runs a post request to send an SMS via the Africa's talking SMS endpoint, this
-  function accepts a map of parameters that should always contain  the `to` address and the `message` to be
-  sent
-
-  ## Parameters
-    - map: a map containing a `to` and `message` key optionally it may also contain `from`, bulk_sms, enqueue, key_word
-      link_id and retry_hours keys, see the docs at https://build.at-labs.io/docs/sms%2Fsending for how to use these keys
-
-  ## Examples
-  iex> AtEx.send_sms(%{to: "+254721978097", message: "Howdy"})
-    {:ok,
-      %{
-        "SMSMessageData" => %{
-        "Message" => "Sent to 1/1 Total Cost: ZAR 0.1124",
-        "Recipients" => [
-        %{
-          "cost" => "KES 0.8000",
-          "messageId" => "ATXid_96e52a761a82c1bad58e885109224aad",
-          "number" => "+254721978097",
-          "status" => "Success",
-          "statusCode" => 101
-        }
-        ]
-      }
-    }}
-  """
-  defdelegate send_sms(map), to: Sms.Bulk
-
-  @doc """
-
-  This function makes a get request to fetch an SMS via the Africa's talking SMS endpoint, this
-  function accepts an map of parameters that optionally accepts `lastReceivedId` of the message.
-  sent
-
-  ## Parameters
-    - map: an empty map or a map containing optionally `lastReceivedId` of the message to be fetched
-
-  ## Examples
-      iex> AtEx.fetch_sms(%{})
-      {:ok,
-        %{"SMSMessageData" =>
-          %{"Messages" =>
-            []
-          }
-        }
-      }
-
-  """
-  defdelegate fetch_sms(map), to: Sms.Bulk
 end
